Prepared for travis
