+++ .travis.yaml
@@ -7,4 +7,4 @@ cache:
 		- node_modules
 script:
 	- npm test
-	- npm run build
+	# - npm run build
